Adds api wrapper, removes legacy code
