--- mantra-game/plugins/binary-bitstream-buffer/example-collection.js
@@ -1,117 +1,0 @@
-import PlayerCodec from './bbb4.js';
-
-const actionTypes = {
-  'gametick': 0,
-  'assign_id': 1,
-  'become_ticker': 2,
-  'pong': 3
-  // ... other action types
-};
-
-// TODO: we'd expect these values to be in the schema as well
-// Enum mapping for player types
-const entityTypes = {
-  'PLAYER': 0,
-  'BULLET': 1,
-  'BLOCK': 2,
-  'BORDER': 3,
-  'BODY': 4
-  // ... other types
-};
-
-const _schema = {
-  id: { type: 'UInt16' },
-  name: { type: 'UTF8String' },
-  type: { type: 'Enum', enum: entityTypes },
-  position: {
-    type: 'Record',
-    schema: {
-      x: { type: 'Int32' },
-      y: { type: 'Int32' }
-    }
-  },
-  velocity: {
-    type: 'Record',
-    schema: {
-      x: { type: 'Int32' },
-      y: { type: 'Int32' }
-    }
-  },
-  width: { type: 'Int32' },
-  height: { type: 'Int32' },
-  rotation: { type: 'Int32' }, // TODO: special case with radians->bytes optimization
-  mass: { type: 'Int32' },
-  health: { type: 'Int32' },
-  depth: { type: 'Float64' },
-  lifetime: { type: 'Int32' },
-  radius: { type: 'Float64' },
-  isSensor: { type: 'Boolean' },
-  isStatic: { type: 'Boolean' },
-  destroyed: { type: 'Boolean' },
-  owner: { type: 'UInt16' },
-  maxSpeed: { type: 'Int32' }
-
-};
-
-const snapshotSchema = {
-  id: { type: 'UInt16' },
-  state: {
-    type: 'Collection',
-    schema: _schema
-  }
-}
-
-let players = [
-  {
-
-    id: 1,
-    name: 'Bunny',
-    type: 'PLAYER',
-    position: { x: 10, y: 20 },
-    velocity: { x: 1, y: 1 },
-    width: 100,
-    height: 100,
-    rotation: 0,
-    mass: 100,
-    health: 100,
-    depth: 10,
-    lifetime: 1000,
-    radius: 100,
-    isSensor: true,
-    isStatic: true,
-    destroyed: true,
-    owner: 0,
-    maxSpeed: 100,
-  },
-  {
-    id: 2,
-    name: 'Turtle',
-    type: 'BLOCK',
-    rotation: 157
-  },
-  {
-    id: 3,
-    name: 'Turtle',
-
-    type: 'PLAYER',
-    //position: { x: 50, y: 60 },
-    //velocity: { x: 3, y: 35 },
-    rotation: 314
-  }
-];
-
-let codec = new PlayerCodec(snapshotSchema, actionTypes);
-
-//console.log('before', players)
-let msg = {
-  id: 123,
-  state: players
-}
-
-let compressed = codec.encodeMessage(msg);
-
-console.log(compressed);
-
-let decompressed = codec.decodeMessage(compressed);
-
-console.log(decompressed);
